Use online streamStatus for Generic collection
Since command collection "Generic" contains command list
`Online.xml`, it makes sense to limit its application only to
online streams. This shouldn't be a functional change, since
command collection "Offline" is defined earlier in the list of
command collections, but it's better not to rely on it anyway
and be more explicit with when command collection "Generic"
is supposed to be used.

So use streamStatus `online` in command collection "Generic".